Small changes to reduce complexity from review
