+++ qcfractal/testing.py
@@ -9,6 +9,7 @@ import socket
 import threading
 import pymongo
 from contextlib import contextmanager
+from collections import Mapping
 
 import pytest
 from tornado.ioloop import IOLoop
@@ -47,6 +48,7 @@ else:
 def has_module(name):
     return _programs[name]
 
+
 # Add a number of module testing options
 using_fireworks = pytest.mark.skipif(has_module('fireworks') is False, reason=_import_message.format('fireworks'))
 using_dask = pytest.mark.skipif(
@@ -58,6 +60,17 @@ using_torsiondrive = pytest.mark.skipif(has_module('torsiondrive') is False, rea
 using_unix = pytest.mark.skipif(os.name.lower() != 'posix', reason='Not on Unix operating system, '
                                                                    'assuming Bash is not present')
 
+
+# More complex than a simple top-level merge {**x, **y} which does not handle nested dict
+def recursive_dict_merge(base_dict, dict_to_merge_in):
+    for k, v in dict_to_merge_in.items():
+        if (k in base_dict and isinstance(base_dict[k], dict)
+                and isinstance(dict_to_merge_in[k], Mapping)):
+            recursive_dict_merge(base_dict[k], dict_to_merge_in[k])
+        else:
+            base_dict[k] = dict_to_merge_in[k]
+
+
 # Check for MongoDB connection
 def check_active_mongo_server():
 
@@ -66,6 +79,7 @@ def check_active_mongo_server():
         client.server_info()
     except:
         pytest.skip("Could not find an activate mongo test instance at 'localhost:27017'.")
+
 
 ### Server testing mechanics
 
@@ -123,6 +137,7 @@ def active_loop(loop):
             thread.join(timeout=5)
         except:
             pass
+
 
 @pytest.fixture(scope="module")
 def test_server(request):
