+++ qcfractal/tests/test_services.py
@@ -4,10 +4,9 @@ Tests the on-node procedures compute capabilities.
 
 from qcfractal import testing
 # Pytest Fixture import
-from qcfractal.testing import dask_server_fixture
+from qcfractal.testing import dask_server_fixture, recursive_dict_merge
 import pytest
 import copy
-from collections import Mapping
 
 import qcfractal.interface as portal
 
@@ -48,32 +47,24 @@ def torsiondrive_fixture(dask_server_fixture):
         instance_options = copy.deepcopy(torsiondrive_options)
         instance_options["torsiondrive_meta"]["grid_spacing"] = [grid_spacing]
 
-        # More complex than a simple top-level merge {**x} which does not handle nested dict
-        def recursive_dict_merge(base_dict, dict_to_merge_in):
-            for k, v in dict_to_merge_in.items():
-                if (k in base_dict and isinstance(base_dict[k], dict)
-                        and isinstance(dict_to_merge_in[k], Mapping)):
-                    recursive_dict_merge(base_dict[k], dict_to_merge_in[k])
-                else:
-                    base_dict[k] = dict_to_merge_in[k]
         # instance_options = {**instance_options, **keyword_augments}
         recursive_dict_merge(instance_options, keyword_augments)
-        return client.add_service("torsiondrive", [mol_ret["hooh"]], instance_options)
+        ret = client.add_service("torsiondrive", [mol_ret["hooh"]], instance_options)
+        nanny = dask_server_fixture.objects["queue_nanny"]
+        nanny.await_services(max_iter=5)
+        assert len(nanny.list_current_tasks()) == 0
+        return ret
 
-    yield spin_up_test, dask_server_fixture, client
+    yield spin_up_test, client
 
 
 def test_service_torsiondrive(torsiondrive_fixture):
     """"Ensure torsiondrive works as intended gives the correct result"""
     # This test does not ensure de-duplication of work,
-    spin_up_test, server_fixture, client = torsiondrive_fixture
+    spin_up_test, client = torsiondrive_fixture
 
     ret = spin_up_test()
     compute_key = ret["submitted"][0]
-
-    nanny = server_fixture.objects["queue_nanny"]
-    nanny.await_services(max_iter=5)
-    assert len(nanny.list_current_tasks()) == 0
 
     # Get a TorsionDriveORM result and check data
     result = client.get_procedures({"procedure": "torsiondrive"})[0]
@@ -88,17 +79,12 @@ def test_service_torsiondrive(torsiondrive_fixture):
 def test_service_torsiondrive_duplicates(torsiondrive_fixture):
     """Ensure that duplicates are properly caught and yield the same results without calculation"""
     # This test does not ensure accuracy, there is another test for that
-    spin_up_test, server_fixture, client = torsiondrive_fixture
+    spin_up_test, client = torsiondrive_fixture
     # Run the test without modifications
     _ = spin_up_test()
-    nanny = server_fixture.objects["queue_nanny"]
-    nanny.await_services(max_iter=5)
-    # Ensure the job finished
-    assert len(nanny.list_current_tasks()) == 0
     # Augment the input for torsion drive to yield a new hash procedure hash,
     # but not a new task set
     _ = spin_up_test(torsiondrive_meta={"meaningless_entry_to_change_hash": "Waffles!"})
-    nanny.await_services(max_iter=5)
     procedures = client.get_procedures({"procedure": "torsiondrive"})
     assert len(procedures) == 2  # Make sure only 2 procedures are yielded
     base_run, duplicate_run = procedures
